# Use bufio.NewReader and add benchmark