This PR adds a benchmark of RecordIO reading. The following command runs unit tests and benchmarks. 

```
go test -bench=.
```

Without `-bench=.`, it runs no benchmark.

This PR also tries to improve reading performance by creating a `bufio.Reader` for each chunk to be read. However, it doesn't seem a significant change. Maybe we should try some other way.

Before using `bufio.Reader`:
```
goos: darwin
goarch: amd64
pkg: github.com/wangkuiyi/recordio
BenchmarkRead/reading_records_00000_to_00050-4         	       5	 293563564 ns/op
BenchmarkRead/reading_records_00050_to_00100-4         	       3	 445585152 ns/op
BenchmarkRead/reading_records_00100_to_00150-4         	       5	 298169020 ns/op
PASS
ok  	github.com/wangkuiyi/recordio	8.330s
```

After using it:

```
goos: darwin
goarch: amd64
pkg: github.com/wangkuiyi/recordio
BenchmarkRead/reading_records_00000_to_00050-4         	       5	 296029121 ns/op
BenchmarkRead/reading_records_00050_to_00100-4         	       3	 455643627 ns/op
BenchmarkRead/reading_records_00100_to_00150-4         	       3	 365613422 ns/op
PASS
ok  	github.com/wangkuiyi/recordio	7.143s
```