@ywskycn 2019-07-18
Tried locally. Using bufio or not, deliver similar performance.